Always print the result, independiently of the verbose mode.
This is because it maybe anoting to check always the return code $? if
no output comes out after invocation.

I don't remember, but I think I disabled it when testing runs from cron.

Maybe in a future I disable it again, but for now I think it's better
this way.